[BW] Update testing infra
